+++ lib/src/test/_index.dart
@@ -3,3 +3,4 @@
 // All rights reserved
 
 export 'echo.dart';
+export 'json_data_service.dart';