+++ lib/src/typedefs.dart
@@ -1,0 +1,9 @@
+// Copyright ©2022, GM Consult (Pty) Ltd
+// BSD 3-Clause License
+// All rights reserved
+
+/// Alias for `Map<String, dynamic>`.
+typedef JSON = Map<String, dynamic>;
+
+/// Alias for `Map<String, Map<String, dynamic>>;`.
+typedef JsonCollection = Map<String, Map<String, dynamic>>;