+++ test/data/data.dart
@@ -1,9 +1,0 @@
-const json = [
-  {'term': 'bodrer', 'other': 'bodrer', 'dL': 0.0, 'cLs': 1.0},
-  {'term': 'bodrer', 'other': 'border', 'dL': 0.0, 'cLs': 1.0},
-  {'term': 'bodrer', 'other': 'boarder', 'dL': 0.222395656, 'cLs': 0.77761454},
-  {'term': 'bodrer', 'other': 'brother', 'dL': 0.22253904, 'cLs': 0.7776176},
-  {'term': 'bodrer', 'other': 'board', 'dL': 0.26303837, 'cLs': 0.736},
-  {'term': 'bodrer', 'other': 'broad', 'dL': 0.263035476, 'cLs': 0.73697546},
-  {'term': 'bodrer', 'other': 'bored', 'dL': 0.2630354, 'cLs': 0.7369766},
-];