+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Copyright (c) 2022, GM Consult Pty Ltd
 All rights reserved. 
 -->
 
+### 0.0.1
+**BETA 1**
 
+- class `Echo` coded and tested.
+- class `JsonDataService` coded and tested.
 
 ### 0.0.1-1